refactor(teams): extract teams doctor data collection for reuse
